+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.2] - 2024-09-25
+
+### Added
+- Configuration file `config/pesawise.php`
+- Updated `PesawiseProvider` to correctly load and publish config
+
+### Fixed
+- Issue with missing configuration file
+
 ## [1.0.1] - 2024-09-25
 
 ### Changed